support time.Duration for tr watch --timeout and tc launch --wait-timeout

+++ cmd/utils.go
@@ -80,7 +80,7 @@ func readOrganisationUIDFromFile() string {
 	return strings.TrimSpace(string(content))
 }
 
-func watchTestRun(testRunUID string, maxWatchTime int) {
+func watchTestRun(testRunUID string, maxWatchTime float64) {
 	client := NewClient()
 
 	started := time.Now()
@@ -103,7 +103,7 @@ func watchTestRun(testRunUID string, maxWatchTime int) {
 			os.Exit(0)
 		}
 
-		if maxWatchTime > 0 && int(runningSince) > maxWatchTime {
+		if int(maxWatchTime) > 0 && int(runningSince) > int(maxWatchTime) {
 			os.Exit(2)
 		}
